restructed the project strcture for MVC
